Add student marks table and Logo module , readme file
Introduced a new feature to input student names and marks, display them in a formatted table, and identify the highest and lowest marks. Added a new Logo.py module for ASCII art used in the CLI, and updated the main menu to include the student table option. You can also copy that table and paste any where. and also can calculate the marks.